requirements.txt
discord.py>=2.4.0
audioop-lts>=0.2.1; python_version >= "3.13"
python-dotenv==1.0.1
anthropic==0.25.0
openai==1.12.0
apscheduler==3.10.4
aiosqlite==0.20.0
requests==2.31.0
aiohttp==3.9.3
supabase==2.4.0
pinecone-client==3.2.2
playwright==1.42.0
pytest==8.1.0
pytest-asyncio==0.23.5
